ref: separate out course prereq and use enum
